==> path_3a/R/step-2_individual-data-reformatting.Rmd <==
---
title: "Path 3.A - Step 2 - Individual data reformatting"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

```

# Load packages

```{r base}

# 1. Required packages ----

library(tidyverse) # Core tidyverse packages
library(readxl) # To read excel files
library(lubridate) # To transform date

# 2. Create a vector of selected variables names to check which ones are missing ----

selected_variables <- c("DatasetID", "Area", "Country", "Archipelago", 
                        "Location", "Site", "Replicate", "Zone", "Latitude", "Longitude", 
                        "Depth", "Year", "Date", "Method", "Observer", "Cover", "Tax_ID")

```

# Data cleaning

## Data contributor 1

```{r DATA1}

# 1. Import data ----

# Main data
data_main <- read_excel("./../data/01_raw/data_contributor_1/dataset1.xlsx", 
                        sheet = "Main data")

# Substrate code data
data_code <- read_excel("./../data/01_raw/data_contributor_1/dataset1.xlsx", 
                        sheet = "Substrate codes")

# 2. Merge and modifications ----

data_main <- data_main %>% 
  left_join(., data_code) %>% # Merge main data with substrates codes
  select(-Season, -Code, -Remarks) %>% # Delete useless variables
  rename(Zone = Habitat, Observer = Oserver, Replicate = Station, 
         Latitude = Lat, Longitude = Long, Tax_ID = Substrate) %>% # Rename variables
  group_by(Year, Date, Observer, Site, Zone, Latitude, Longitude, Replicate, Tax_ID) %>% 
  summarise(Cover = n()) %>% # Get number of point of each substrate type on the transect
  ungroup() %>% 
  mutate(Cover = Cover*2, # Transform Cover to percentage of cover
         Latitude = as.numeric(Latitude), # Convert Latitude to numeric
         Longitude = as.numeric(Longitude), # Convert Longitude to numeric
         DatasetID = "DATA1",
         Area = "Pacific",
         Country = "USA",
         Archipelago = "Mariana Islands",
         Location = "Guam",
         Method = "Point Intercept Transect") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA1.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_code)

```

## Data contributor 2

```{r DATA2}

# 1. Import data ----

# Main data
data_main <- read_excel("./../data/01_raw/data_contributor_2/dataset2.xlsx", 
                        sheet = "Transect")

# Site data
data_site <- read_excel("./../data/01_raw/data_contributor_2/dataset2.xlsx", 
                        sheet = "Informations")

# 2. Merge and modifications ----

data_main <- data_main %>% 
  left_join(., data_site) %>% # Merge main data and substrates codes
  select(-Total) %>% # Delete useless variables
  rename(Replicate = Transect, Latitude = Lat, Longitude = Long) %>% # Rename variables
  pivot_longer("Non living":"Turf algae", names_to = "Tax_ID", values_to = "Cover") %>% # Pivot the data (wide to long format)
  filter(Cover != 0) %>% # Remove rows where cover == 0 (useless)
  mutate(Lat_degree = as.numeric(str_split_fixed(Latitude, "°|'|\"", 4)[,1]),
         Lat_min = as.numeric(str_split_fixed(Latitude, "°|'|\"", 4)[,2]),
         Lat_sec = as.numeric(str_split_fixed(Latitude, "°|'|\"", 4)[,3]),
         Lat_side = str_split_fixed(Latitude, "°|'|\"", 4)[,4],
         Latitude = Lat_degree + (Lat_min/60) + (Lat_sec/3600),
         Latitude = ifelse(Lat_side == "N", Latitude, ifelse(Lat_side == "S", -Latitude, NA))) %>% # Convert Lat. coord.*
  select(-Lat_degree, -Lat_min, -Lat_sec, -Lat_side) %>% # Delete useless variables
  mutate(Long_degree = as.numeric(str_split_fixed(Longitude, "°|'|\"", 4)[,1]),
         Long_min = as.numeric(str_split_fixed(Longitude, "°|'|\"", 4)[,2]),
         Long_sec = as.numeric(str_split_fixed(Longitude, "°|'|\"", 4)[,3]),
         Long_side = str_split_fixed(Longitude, "°|'|\"", 4)[,4],
         Longitude = Long_degree + (Long_min/60) + (Long_sec/3600),
         Longitude = ifelse(Long_side == "E", Longitude, ifelse(Long_side == "W", -Longitude, NA))) %>% # Convert Long. coord.*
  select(-Long_degree, -Long_min, -Long_sec, -Long_side) %>% # Delete useless variables
  mutate(Replicate = str_sub(Replicate, start = 3, end = 3), # Extract the number of the replicate
         DatasetID = "DATA2",
         Area = "Caribbean") # Add variables absent in data (can be obtained through data contributor)
  
# * functions can be used for this, but this method is more polyvalent

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA2.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site)

```

## Data contributor 3

```{r DATA3}

# 1. Import data ----

# Import first sheet

data_main <- read_excel("./../data/01_raw/data_contributor_3/dataset3.xlsx", sheet = 1) %>% 
  mutate(Site = excel_sheets("./../data/01_raw/data_contributor_3/dataset3.xlsx")[1])
  
# Make a loop to bind all sheets

for(i in 2:length(excel_sheets("./../data/01_raw/data_contributor_3/dataset3.xlsx"))){
  
  data_main <- read_excel("./../data/01_raw/data_contributor_3/dataset3.xlsx", sheet = i) %>% 
    mutate(Site = excel_sheets("./../data/01_raw/data_contributor_3/dataset3.xlsx")[i]) %>% 
    bind_rows(data_main, .)
  
}

# 2. Merge and modifications ----

data_main <- data_main %>% 
  select(-"AllCorals":-"SUM_ALL") %>% # Delete useless variables (here, those corresponding to the sum of oth. var.)
  pivot_longer("Acanthastrea":"Wand", names_to = "Tax_ID", values_to = "Cover") %>% # Pivot the data (wide to long format)
  filter(Cover != 0) %>% # Remove rows where cover == 0 (useless)
  rename(Replicate = Transect, Zone = Habitat) %>% # Rename variables
  mutate(Latitude = as.numeric(Latitude), # Convert Latitude to numeric
         Longitude = as.numeric(Longitude), # Convert Longitude to numeric
         DatasetID = "DATA3",
         Area = "Australia",
         Country = "Australia",
         Location = "Lizard Island") # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA3.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, i)

```

## Data contributor 4

```{r DATA4}

# 1. Import data ----

# Main data 1
data_main_1 <- read.csv("./../data/01_raw/data_contributor_4/dataset4a.csv", sep = ";")

# Main data 2
data_main_2 <- read.csv("./../data/01_raw/data_contributor_4/dataset4b.csv", sep = ";")

# Substrate code
data_code <- read.csv("./../data/01_raw/data_contributor_4/dataset4c.csv", sep = ";")

# 2. Merge and modifications ----

data_main <- data_main_1 %>% 
  bind_rows(., data_main_2) %>% # Bind the two dataframes
  pivot_longer("CAB":"BC", names_to = "Code", values_to = "Cover") %>% # Pivot the data (wide to long format)
  filter(Cover != 0) %>% # Remove rows where cover == 0 (useless)
  left_join(., data_code) %>% # Merge main data with substrates codes
  select(-Sum, -HCC, -Code) %>% # Delete useless variables
  rename(Latitude = Lat, Longitude = Long, Replicate = Transect, Tax_ID = Category) %>% # Rename variables
  mutate(DatasetID = "DATA4",
         Area = "Indian Ocean",
         Country = "Maldives",
         Archipelago = "Maldives",
         Date = format(strptime(as.character(Date), "%d/%m/%Y"), "%Y-%m-%d"),
         Zone = str_replace_all(Zone, c("OS" = "Outer Slope",
                                        "LG" = "Lagoon"))) # Add variables absent in data (can be obtained through data contributor)

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA4.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_main_1, data_main_2, data_code)

```

## Data contributor 5

```{r DATA5.1}

# 1. Import data ----

# Main data
data_main <- read_excel("./../data/01_raw/data_contributor_5/dataset5a.xlsx", sheet = "Sheet1")

# Site data
data_site <- read_excel("./../data/01_raw/data_contributor_5/dataset5c.xlsx", sheet = "Coordinates")

# 2. Merge and modifications ----

data_main <- data_main %>% 
  select(-Total) %>% 
  pivot_longer("ACANTHASTREA":"TURBINARIA", names_to = "Tax_ID", values_to = "Cover") %>% 
  filter(Cover != 0) %>% 
  left_join(., data_site) %>% 
  rename(Year = YEAR, Replicate = "Site No", Latitude = Lat, Longitude = Long) %>% 
  mutate(DatasetID = "DATA5.1",
         Area = "Caribbean",
         Country = "Panama",
         Archipelago = "Bocas del Toro",
         Latitude = as.numeric(Latitude), # Convert Latitude to numeric
         Longitude = as.numeric(Longitude), # Convert Longitude to numeric
         Depth = Depth*0.3048) # Convert depth in feet to depth in meters

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA5.1.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site)

```

```{r DATA5.2}

# 1. Import data ----

# Main data
data_main <- read_excel("./../data/01_raw/data_contributor_5/dataset5b.xlsx", sheet = "Sheet1")

# Site data
data_site <- read_excel("./../data/01_raw/data_contributor_5/dataset5c.xlsx", sheet = "Coordinates")

# 2. Merge and modifications ----

data_main <- data_main %>% 
  pivot_longer("Abiotic":"Algae_other", names_to = "Tax_ID", values_to = "Cover") %>% 
  filter(Cover != 0) %>% 
  left_join(., data_site) %>% 
  rename(Latitude = Lat, Longitude = Long) %>% 
  mutate(DatasetID = "DATA5.2",
         Area = "Caribbean",
         Country = "Panama",
         Archipelago = "Bocas del Toro",
         Latitude = as.numeric(Latitude), # Convert Latitude to numeric
         Longitude = as.numeric(Longitude), # Convert Longitude to numeric
         Depth = Depth*0.3048) # Convert depth in feet to depth in meters

# 3. Export data ----

write.csv2(data_main, "./../data/02_reformatted/02-reformatted_DATA5.2.csv", row.names = FALSE)

# 4. Missing variables ----

setdiff(selected_variables, colnames(data_main)) # Get names of missing variables

# 5. Remove useless datasets and variables ----

rm(data_main, data_site)

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Jeremy WICQUART | jeremywicquart@gmail.com | `r format(Sys.time())`
